Use function style cast and remove unnecesary cast

+++ source/blender/nodes/function/nodes/node_fn_random_value.cc
@@ -171,8 +171,7 @@ static void node_build_multi_function(NodeMultiFunctionBuilder &builder)
               std::swap(min_value, max_value);
             }
             const uint32_t hash = noise::hash(id, seed);
-            const uint64_t range = static_cast<int64_t>(max_value) -
-                                   static_cast<int64_t>(min_value) + 1;
+            const uint64_t range = int64_t(max_value) - min_value + 1;
             return min_value + hash % (range);
           },
           mf::build::exec_presets::SomeSpanOrSingle<2>());
